+++ VisitorApp/TestWindow.xaml
@@ -4,10 +4,12 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:CampingApplication.VisitorApp"
-        xmlns:views="clr-namespace:CampingApplication.VisitorApp.Views"
+        xmlns:views="clr-namespace:CampingApplication.VisitorApp.Views.Booking"
         mc:Ignorable="d"
         Title="TestWindow" Background="LightGray" Height="450" Width="800">
     <Grid VerticalAlignment="Center">
-        <views:BookingView />
+        <Grid x:Name="ContentWindow">
+
+        </Grid>
     </Grid>
 </Window>
