+++ VisitorApp/Views/Booking/BookingDetailsView.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -23,6 +24,22 @@ namespace CampingApplication.VisitorApp.Views.Booking
         public BookingDetailsView()
         {
             InitializeComponent();
+        }
+
+        public void SetDetails(int id, DateTime startDate, DateTime endDate, int amountOfNights, float pricePerNight, float totalPrice)
+        {
+            ID.Text = id.ToString();
+            StartDate.Text = startDate.ToShortDateString();
+            EndDate.Text = endDate.ToShortDateString();
+            AmountOfNights.Text = amountOfNights.ToString();
+            PricePerNight.Text = FormatPrice(pricePerNight);
+            TotalPrice.Text = FormatPrice(totalPrice);
+        }
+
+        private static string FormatPrice(float priceInEuros)
+        {
+            var cultureInfo = new CultureInfo("nl-NL");
+            return priceInEuros.ToString("C", cultureInfo);
         }
     }
 }
