+++ VisitorApp/Views/BookingView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="CampingApplication.VisitorApp.Views.BookingView"
+﻿<UserControl x:Class="CampingApplication.VisitorApp.Views.Booking.BookingView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -41,9 +41,15 @@
 
     <Border Padding="15" Background="White" HorizontalAlignment="Center" VerticalAlignment="Center" CornerRadius="10">
         <StackPanel>
-            <TextBlock Margin="0,0,0,10" FontSize="30" FontWeight="Bold">Boeking Voltooien</TextBlock>
+            <StackPanel Margin="0,0,0,10" Orientation="Horizontal">
+                <Viewbox VerticalAlignment="Center" Width="32" Height="32">
+                    <ContentControl Foreground="Black" Content="{DynamicResource ArrowLeft}">
+                    </ContentControl>
+                </Viewbox>
+                <TextBlock LineHeight="45" Margin="20,0,0,0" FontSize="30" FontWeight="Bold"><Run Text="Boeking Voltooien"/></TextBlock>
+            </StackPanel>
             <StackPanel Orientation="Horizontal">
-                <local:BookingDetailsView HorizontalAlignment="Left" VerticalAlignment="Top"/>
+                <local:BookingDetailsView x:Name="Details" HorizontalAlignment="Left" VerticalAlignment="Top"/>
 
                 <Border Margin="10,0,0,0"  Style="{StaticResource InformationBorderStyle}">
                     <StackPanel>
